+++ syncany-lib/src/main/java/org/syncany/plugins/transfer/TransferPluginOptions.java
@@ -19,12 +19,9 @@ package org.syncany.plugins.transfer;
 
 import java.lang.reflect.Field;
 import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import org.simpleframework.xml.Element;
 import org.syncany.util.ReflectionUtil;
-import org.syncany.util.StringUtil;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
@@ -38,7 +35,6 @@ import com.google.common.primitives.Ints;
  * @author Christian Roth <christian.roth@port17.de>
  */
 public class TransferPluginOptions {
-	private static final Logger logger = Logger.getLogger(TransferPluginOptions.class.getName());
 	private static final int MAX_NESTED_LEVELS = 3;
 
 	/**
@@ -64,7 +60,7 @@ public class TransferPluginOptions {
 		return options.build();
 	}
 
-	private static TransferPluginOption getOptionFromField(Field field, Class<? extends TransferSettings> transferSettingsClass, int level) {		
+	private static TransferPluginOption getOptionFromField(Field field, Class<? extends TransferSettings> transferSettingsClass, int level) {
 		Element elementAnnotation = field.getAnnotation(Element.class);
 		Setup setupAnnotation = field.getAnnotation(Setup.class);
 
@@ -88,10 +84,10 @@ public class TransferPluginOptions {
 		boolean isNestedOption = TransferSettings.class.isAssignableFrom(field.getType());
 
 		if (isNestedOption) {
-			return createNestedOption(field, level, name, description, fileType, encrypted, sensitive, singular, visible, required, callback, converter);			
+			return createNestedOption(field, level, name, description, fileType, encrypted, sensitive, singular, visible, required, callback, converter);
 		}
 		else {
-			return createNormalOption(field, transferSettingsClass, name, description, fileType, encrypted, sensitive, singular, visible, required, callback, converter);			
+			return createNormalOption(field, transferSettingsClass, name, description, fileType, encrypted, sensitive, singular, visible, required, callback, converter);
 		}
 	}
 
@@ -99,7 +95,7 @@ public class TransferPluginOptions {
 	private static TransferPluginOption createNestedOption(Field field, int level, String name, String description, FileType fileType,
 			boolean encrypted, boolean sensitive, boolean singular, boolean visible, boolean required,
 			Class<? extends TransferPluginOptionCallback> callback, Class<? extends TransferPluginOptionConverter> converter) {
-		
+
 		if (++level > MAX_NESTED_LEVELS) {
 			throw new RuntimeException("Plugin uses too many nested transfer settings (max allowed value: " + MAX_NESTED_LEVELS + ")");
 		}
@@ -112,24 +108,8 @@ public class TransferPluginOptions {
 	private static TransferPluginOption createNormalOption(Field field, Class<? extends TransferSettings> transferSettingsClass, String name,
 			String description, FileType fileType, boolean encrypted, boolean sensitive, boolean singular, boolean visible, boolean required,
 			Class<? extends TransferPluginOptionCallback> callback, Class<? extends TransferPluginOptionConverter> converter) {
-		
-		if (Enum.class.isAssignableFrom(field.getType())) {
-			Object[] enumValues = getEnumValues(field, transferSettingsClass);			
-			description = description + " (Valid values are: " + StringUtil.join(enumValues, ", ") + ")";
-		}
 
 		return new TransferPluginOption(field, name, description, field.getType(), fileType, encrypted, sensitive, singular, visible, required, callback, converter);
-	}
-
-	private static Object[] getEnumValues(Field field, Class<? extends TransferSettings> transferSettingsClass) {
-		Object[] enumValues = field.getType().getEnumConstants();		
-		logger.log(Level.FINE, "Enum values are: " + StringUtil.join(enumValues, ", "));
-
-		if (enumValues == null) {
-			throw new RuntimeException("Invalid TransferSettings class found: Enum at " + transferSettingsClass + " has no values");
-		}
-		
-		return enumValues;
 	}
 
 	private static List<Field> getOrderedFields(Class<? extends TransferSettings> transferSettingsClass) {